+++ CHANGELOG.md
@@ -8,6 +8,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-10-16
+
 ### ⚠️ Breaking changes
 
 - Validation accessors are generated only for public properties. ([#11](https://github.com/nesk/akkurate/issues/11) [#15](https://github.com/nesk/akkurate/issues/15))
@@ -65,7 +67,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 First release
 
-[Unreleased]: https://github.com/nesk/akkurate/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/nesk/akkurate/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/nesk/akkurate/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/nesk/akkurate/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/nesk/akkurate/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/nesk/akkurate/releases/tag/0.1.0